+++ exchangelib/ewsdatetime.py
@@ -1,13 +1,16 @@
 import datetime
 import logging
+import warnings
 
 import dateutil.parser
-import pytz
-import pytz.exceptions
+try:
+    import zoneinfo
+except ImportError:
+    from backports import zoneinfo
 import tzlocal
 
-from .errors import NaiveDateTimeNotAllowed, UnknownTimeZone, AmbiguousTimeError, NonExistentTimeError
-from .winzone import PYTZ_TO_MS_TIMEZONE_MAP, MS_TIMEZONE_TO_PYTZ_MAP
+from .errors import NaiveDateTimeNotAllowed, UnknownTimeZone
+from .winzone import IANA_TO_MS_TIMEZONE_MAP, MS_TIMEZONE_TO_IANA_MAP
 
 log = logging.getLogger(__name__)
 
@@ -79,13 +82,6 @@ class EWSDateTime(datetime.datetime):
 
     def __new__(cls, *args, **kwargs):
         # pylint: disable=arguments-differ
-        # Not all Python versions have the same signature for datetime.datetime
-        """
-        Inherits datetime and adds extra formatting required by EWS. Do not set tzinfo directly. Use
-        EWSTimeZone.localize() instead.
-        """
-        # We can't use the exact signature of datetime.datetime because we get pickle errors, and implementing pickle
-        # support requires copy-pasting lots of code from datetime.datetime.
         if not isinstance(kwargs.get('tzinfo'), (EWSTimeZone, type(None))):
             raise ValueError('tzinfo must be an EWSTimeZone instance')
         return super().__new__(cls, *args, **kwargs)
@@ -97,8 +93,8 @@ class EWSDateTime(datetime.datetime):
 
         """
         if not self.tzinfo:
-            raise ValueError('EWSDateTime must be timezone-aware')
-        if self.tzinfo.zone == 'UTC':
+            raise ValueError('%r must be timezone-aware' % self)
+        if self.tzinfo.key == 'UTC':
             return self.strftime('%Y-%m-%dT%H:%M:%SZ')
         return self.replace(microsecond=0).isoformat()
 
@@ -111,7 +107,7 @@ class EWSDateTime(datetime.datetime):
         elif isinstance(d.tzinfo, EWSTimeZone):
             tz = d.tzinfo
         else:
-            tz = EWSTimeZone.from_pytz(d.tzinfo)
+            tz = EWSTimeZone(d.tzinfo.key)
         return cls(d.year, d.month, d.day, d.hour, d.minute, d.second, d.microsecond, tzinfo=tz)
 
     def astimezone(self, tz=None):
@@ -145,8 +141,7 @@ class EWSDateTime(datetime.datetime):
         # Parses several common datetime formats and returns timezone-aware EWSDateTime objects
         if date_string.endswith('Z'):
             # UTC datetime
-            naive_dt = super().strptime(date_string, '%Y-%m-%dT%H:%M:%SZ')
-            return UTC.localize(naive_dt)
+            return super().strptime(date_string, '%Y-%m-%dT%H:%M:%SZ').replace(tzinfo=UTC)
         if len(date_string) == 19:
             # This is probably a naive datetime. Don't allow this, but signal caller with an appropriate error
             local_dt = super().strptime(date_string, '%Y-%m-%dT%H:%M:%S')
@@ -155,7 +150,7 @@ class EWSDateTime(datetime.datetime):
         # strptime '%z' directive cannot yet handle full ISO8601 formatted timezone information (see
         # http://bugs.python.org/issue15873). Use the 'dateutil' package instead.
         aware_dt = dateutil.parser.parse(date_string)
-        return cls.from_datetime(aware_dt.astimezone(UTC))  # We want to return EWSDateTime objects
+        return aware_dt.astimezone(UTC)
 
     @classmethod
     def fromtimestamp(cls, t, tz=None):
@@ -192,109 +187,95 @@ class EWSDateTime(datetime.datetime):
         return EWSDate.from_date(d)  # We want to return EWSDate objects
 
 
-class EWSTimeZone:
+class EWSTimeZone(zoneinfo.ZoneInfo):
     """Represents a timezone as expected by the EWS TimezoneContext / TimezoneDefinition XML element, and returned by
     services.GetServerTimeZones.
 
     """
-    PYTZ_TO_MS_MAP = PYTZ_TO_MS_TIMEZONE_MAP
-    MS_TO_PYTZ_MAP = MS_TIMEZONE_TO_PYTZ_MAP
+    IANA_TO_MS_MAP = IANA_TO_MS_TIMEZONE_MAP
+    MS_TO_IANA_MAP = MS_TIMEZONE_TO_IANA_MAP
 
-    def __eq__(self, other):
-        # Microsoft timezones are less granular than pytz, so an EWSTimeZone created from 'Europe/Copenhagen' may return
-        # from the server as 'Europe/Copenhagen'. We're catering for Microsoft here, so base equality on the Microsoft
-        # timezone ID.
-        if not hasattr(other, 'ms_id'):
-            # Due to the type magic in from_pytz(), we cannot use isinstance() here
-            return NotImplemented
-        return self.ms_id == other.ms_id
-
-    def __hash__(self):
-        # We're shuffling around with base classes in from_pytz(). Make sure we have __hash__() implementation.
-        return super().__hash__()
-
-    @classmethod
-    def from_ms_id(cls, ms_id):
-        # Create a timezone instance from a Microsoft timezone ID. This is lossy because there is not a 1:1 translation
-        # from MS timezone ID to pytz timezone.
+    def __new__(cls, *args, **kwargs):
         try:
-            return cls.timezone(cls.MS_TO_PYTZ_MAP[ms_id])
-        except KeyError:
-            if '/' in ms_id:
-                # EWS sometimes returns an ID that has a region/location format, e.g. 'Europe/Copenhagen'. Try the
-                # string unaltered.
-                return cls.timezone(ms_id)
-            raise UnknownTimeZone("Windows timezone ID '%s' is unknown by CLDR" % ms_id)
-
-    @classmethod
-    def from_pytz(cls, tz):
-        # pytz timezones are dynamically generated. Subclass the tz.__class__ and add the extra Microsoft timezone
-        # labels we need.
-
-        # type() does not allow duplicate base classes. For static timezones, 'cls' and 'tz' are the same class.
-        base_classes = (cls,) if cls == tz.__class__ else (cls, tz.__class__)
-        self_cls = type(cls.__name__, base_classes, dict(tz.__class__.__dict__))
+            instance = super().__new__(cls, *args, **kwargs)
+        except zoneinfo.ZoneInfoNotFoundError as e:
+            raise UnknownTimeZone(e.args[0])
         try:
-            self_cls.ms_id = cls.PYTZ_TO_MS_MAP[tz.zone][0]
+            instance.ms_id = cls.IANA_TO_MS_MAP[instance.key][0]
         except KeyError:
-            raise UnknownTimeZone('No Windows timezone name found for timezone "%s"' % tz.zone)
+            raise UnknownTimeZone('No Windows timezone name found for timezone "%s"' % instance.key)
 
         # We don't need the Windows long-format timezone name in long format. It's used in timezone XML elements, but
         # EWS happily accepts empty strings. For a full list of timezones supported by the target server, including
         # long-format names, see output of services.GetServerTimeZones(account.protocol).call()
-        self_cls.ms_name = ''
+        instance.ms_name = ''
+        return instance
 
-        self = self_cls()
-        for k, v in tz.__dict__.items():
-            setattr(self, k, v)
-        return self
+    def __eq__(self, other):
+        # Microsoft timezones are less granular than IANA, so an EWSTimeZone created from 'Europe/Copenhagen' may return
+        # from the server as 'Europe/Copenhagen'. We're catering for Microsoft here, so base equality on the Microsoft
+        # timezone ID.
+        if not isinstance(other, self.__class__):
+            return NotImplemented
+        return self.ms_id == other.ms_id
+
+    @classmethod
+    def from_ms_id(cls, ms_id):
+        # Create a timezone instance from a Microsoft timezone ID. This is lossy because there is not a 1:1 translation
+        # from MS timezone ID to IANA timezone.
+        try:
+            return cls(cls.MS_TO_IANA_MAP[ms_id])
+        except KeyError:
+            if '/' in ms_id:
+                # EWS sometimes returns an ID that has a region/location format, e.g. 'Europe/Copenhagen'. Try the
+                # string unaltered.
+                return cls(ms_id)
+            raise UnknownTimeZone("Windows timezone ID '%s' is unknown by CLDR" % ms_id)
+
+    @classmethod
+    def from_pytz(cls, tz):
+        return cls(tz.zone)
+
+    @classmethod
+    def from_dateutil(cls, tz):
+        key = '/'.join(tz._filename.split('/')[-2:])
+        return cls(key)
 
     @classmethod
     def localzone(cls):
         try:
             tz = tzlocal.get_localzone()
-        except pytz.exceptions.UnknownTimeZoneError:
+        except zoneinfo.ZoneInfoNotFoundError:
+            # Older versions of tzlocal will raise a pytz exception. Let's not depend on pytz just for that.
             raise UnknownTimeZone("Failed to guess local timezone")
-        return cls.from_pytz(tz)
+        # Handle both old and new versions of tzlocal that may return pytz or zoneinfo objects, respectively
+        return cls(tz.key if hasattr(tz, 'key') else tz.zone)
 
     @classmethod
     def timezone(cls, location):
-        # Like pytz.timezone() but returning EWSTimeZone instances
-        try:
-            tz = pytz.timezone(location)
-        except pytz.exceptions.UnknownTimeZoneError:
-            raise UnknownTimeZone("Timezone '%s' is unknown by pytz" % location)
-        return cls.from_pytz(tz)
+        warnings.warn('replace EWSTimeZone.timezone() with just EWSTimeZone()', DeprecationWarning, stacklevel=2)
+        return cls(location)
 
     def normalize(self, dt, is_dst=False):
-        return self._localize_or_normalize(func='normalize', dt=dt, is_dst=is_dst)
+        warnings.warn('normalization is now handled gracefully', DeprecationWarning, stacklevel=2)
+        return dt
 
     def localize(self, dt, is_dst=False):
-        return self._localize_or_normalize(func='localize', dt=dt, is_dst=is_dst)
-
-    def _localize_or_normalize(self, func, dt, is_dst=False):
-        """localize() and normalize() have common code paths
-
-        Args:
-          func:
-          dt:
-          is_dst:  (Default value = False)
-
-        """
-        # super() returns a dt.tzinfo of class pytz.tzinfo.FooBar. We need to return type EWSTimeZone
-        if is_dst is not False:
-            # Not all pytz timezones support 'is_dst' argument. Only pass it on if it's set explicitly.
-            try:
-                res = getattr(super(EWSTimeZone, self), func)(dt, is_dst=is_dst)
-            except pytz.exceptions.AmbiguousTimeError as exc:
-                raise AmbiguousTimeError(str(dt)) from exc
-            except pytz.exceptions.NonExistentTimeError as exc:
-                raise NonExistentTimeError(str(dt)) from exc
-        else:
-            res = getattr(super(EWSTimeZone, self), func)(dt)
-        if not isinstance(res.tzinfo, EWSTimeZone):
-            return res.replace(tzinfo=self.from_pytz(res.tzinfo))
-        return res
+        warnings.warn('replace localize() with dt.replace(tzinfo=tz)', DeprecationWarning, stacklevel=2)
+        if dt.tzinfo is not None:
+            raise ValueError('%r must be timezone-unaware' % dt)
+        dt = dt.replace(tzinfo=self)
+        if is_dst is not None:
+            # DST dates are assumed to always be after non-DST dates
+            dt_before = dt.replace(fold=0)
+            dt_after = dt.replace(fold=1)
+            dst_before = dt_before.dst()
+            dst_after = dt_after.dst()
+            if dst_before > dst_after:
+                dt = dt_before if is_dst else dt_after
+            elif dst_before < dst_after:
+                dt = dt_after if is_dst else dt_before
+        return dt
 
     def fromutc(self, dt):
         t = super().fromutc(dt)
@@ -303,5 +284,5 @@ class EWSTimeZone:
         return EWSDateTime.from_datetime(t)  # We want to return EWSDateTime objects
 
 
-UTC = EWSTimeZone.timezone('UTC')
+UTC = EWSTimeZone('UTC')
 UTC_NOW = lambda: EWSDateTime.now(tz=UTC)  # noqa: E731